FIX Fix remaining 1D tests (#63)
* TST Fix differentiability 1D test

When sending an autograd variable to the FFT wrappers, we need to
make sure to call `detach()` before `numpy()`.

* ENH Remove warnings from ModulusStable.forward()

Specifically, replace the deprecated `ctx.saved_variables` with
`ctx.saved_tensors`.

* BUG Fix 1D transform for batch size more than one

Assignment dimensions are not compatible when batch size is greater
than one. Need to apply `squeeze(dim=1)` before assigning to the
output Tensor.