ENH Index filters by scalars in 1D scattering
Previously, filters (and consequently scattering coefficients)
were indexed using pairs `(j, n)`, where `j` is the subsampling
power (i.e., dyadic scale) and `n` is a linear index running from
`0` to the number of filters minus one. Scattering coefficients
(in the unvectorized output) were similarly indexed, with second-
order coefficients indexed by a four-tuple `(j1, n1, j2, n2)`.

This scheme complicates accessing the filters, since the caller
needs to know in advance what dyadic scale a particular filter
is located in to access it. Indexing the scattering coefficients
poses a similar difficulty.

Instead, filters are now indexed by a single number `n`, with the
dyadic scale stored as an entry in that filter's dictionary (just
like the parameters `xi`, and `sigma`). So the `j` of a previous
`psi1_f[(j, n)]` is now accessed as `psi1_f[n]['j']`. This implies
that `j` is a function of `n`, which makes sense.

Scattering coefficients are indexed similarly, with the previous
index `(j1, n1)`, now becoming `(n1,)` and `(j1, n1, j2, n2)`
becoming `(n1, n2)`. The zeroth-order coefficient is accessed by
`()` (this was previously `0`) to conform with the above pattern.