ENH Replace meta dictionaries with arrays in 1D
Instead of the `compute_meta_scattering()` function returning
a dictionary with keys corresponding to tuples indexing the
unvectorized scattering transform output, we replace this with
a dictionary containing fields `'order`', `'j'`, and so on, each
of which is a Tensor. Missing values are filled in with NaNs.

The advantage of this is that it makes indexing easier, since we
don't have to do list comprehensions to extract relevant data, such
as orders.