ENH Add stub skcuda backend for 3D
Allows us to import the 3D backend when `KYMATIO_BACKEND` is equal to
`skcuda` without throwing an error. However, if anyone tries to use
the backend, it raises a `NotImplementedError`. The `Scattering3D` class
has also been updated to provide a more informative error if the user
tries to create an object when this backend is enabled.

For now, the tests will simply issue a warning under these
circumstances.