REF Make 1D input shapes `(B, T)`, not `(B, 1, T)` (#215)
Previously, inputs to the scattering transform had to have a
channel axis that was always of dimension one. This is no longer
the case. An input of shape `(B, T)` is converted to a scattering
output of shape `(B, C, T1)`, where `C` is the number of scattering
coefficients and `T1` is the subsampled length of the signal.