FIX Make `Scattering3D` work for `J != 2`
Fixes #216 by concatenating orders along scale axis instead of spatial
axis. Also adds a test to catch these larger scales (J) in 3D and fixes
the sample scattering test to account for this new layout. This is
because the ordering new ordering doesn't correspond to the reference
output stored in the `test_data_3d.pt` file, so we need to rearrange the
output of `Scattering3D` to correspond and make the test pass.