API Enable CPU/CUDA "switch" in `Scattering3D`
In other words, give the class the same `cpu()` and `cuda()` methods as
for the 1D and 2D classes. The difference here is that no filters are
actually transferred, it just changes an internal flag. This harmonizes
the interfaces, however, and allow us to potentially optimize things
down the road by moving certain filters onto the GPU when possible.