FIX The codecov bash uploader
The codecov uploader looks for a `coverage.xml` file in the main
directory (the $WORKSPACE environment variable in the Jenkinsfile). If
it can't find one, it tries to call `coverage xml` which converts the
Python coverage file `.coverage` into `coverage.xml`. For some reason,
the uploader was able to do this conversion automatically before, but
now (probably because `coverage` is only accessible through `python3 -m
coverage` now) it fails. The fix is therefore to manually call `python3
-m coverage xml` after running the tests.