TST Make 1D meta filter work again
To slice torch arrays, we need to be using torch arrays as indices. Since
`meta()` now returns numpy arrays, using them to slice torch arrays has a
different behavior (in certain versions of torch). To make sure we get the
desired behavior, we have to cast to torch before slicing.