API Introduce 2D frontend (torch)
So far, only for PyTorch. Note that this commit doesn't change any of the unit
tests to keep things simple. As a result, this breaks the 2D tests that rely on
the old backend framework being present.

We also remove backend-specific 2D tests. This will make the 2D tests
run for the new frontend, since the Scattering2D interface hasn't
changed. We keep these for now just to make sure we're backwards
compatible (since the default frontend is still the PyTorch one), but
once we're comfortable that the new, frontend-specific tests cover this,
we should remove these since they are redundant.