ENH Avoid copying in core 2D scattering code
The old behavior (pre-frontend) was to allocate an output array and fill
it up gradually with scattering coefficients as these were calculated.
With the frontend interface, this was replaced by construction of lists
of arrays that were concatenated at the end, resulting in an extra copy
operation that was not present before. As a result, performance dropped
by about 10%. This commit restores the old (pre-allocating) behavior.

Without the concatenation, we also need to explicitly code for batch
agnosticism by detecting the batch shape and reshaping the arrays. Since
this is a frontend-specific operation (and can be performed
independently of the core scattering code), it is done in the torch
frontend.

The `concatenate` backend primitive is no longer necessary and is
therefore removed.