REF Rename bare Scattering to ScatteringEntry
This way, there's less confusing between the “aliased” `Scattering*D`
classes (that are actually frontends) and the `ScatteringEntry*D`
classes (which just provide an entry point and morphs into one of those
frontends when instantiated). With this commit, *all* references to
`Scattering*D` refer to aliases, which should make development more
straightforward.