ENH Fix for multi-GPU scattering
Filters are registered as buffers in Torch, which means that they get
copied to different GPUs in the multi-GPU setting. However, during
`load_filters`, we assign these filters to the dictionaries `self.phi`
and `self.psi`, which is shared between GPUs. By separating these
dictionaries, we avoid using the wrong filters when computing the
scattering transforms.