revert
From @janden

Hmm so according to https://download.pytorch.org/whl/torch_stable.html, torchvision==0.11.3 is available on py310, but not torchvision==0.11.3+cpu. Looks like the oldest torchvision that runs on py310 and CPU is torchvision==0.12.0+cpu.

Looks like the issue is the same for torch==1.10.2 (CPU version is not available on py310, only GPU). So let's forget about this. Go back to what you had before.