DOC Fix 1D classification examples
Since the datasets module is no longer part of the Kymatio package, we
can't access it when running the scripts as part of a Sphinx Gallery.
To solve this, we instead use the `deeplake` package, which provides a
dataloader for FSDD. Since `fetch_fsdd` is no longer used, we remove
this from `datasets`.