DOC Fix docstrings (#1036)
* DOC Set sphinx toctree max depth

Looks like the default changed between the last few Sphinx versions.
Should match what we had previously.

* DOC Add docstring for `stride` in 1D

Also add deprecation warning to `oversampling`

* DOC Fix indentation, whitespace in docstrings

* DOC Purge documentation for old arguments

I.e., `vectorize` and `average`.

* DOC Make sure `scattering` docstrings are included

Since these are sometimes inherited from the base frontend, they are not
always included on the codereference page with their docstrings. There
are two ways around this. The first is to use `:inherited-members:`, but
this will force us to include a lot of methods from `torch.nn.Module`
and so on that are not relevant to the package. The second approach
(adopted here) is to automatically define a `scattering` method if one
is not present that simply calls the super method.